Refactor innostore backend module to use new backend API.
Fixes: az710

This change will bring the innostore backend into sync with the rest
of the backends with regard to the API including support for
asynchronous folds and the delete/4 function.

Additionally, the test suite has been replaced with a suite of tests
that are similar to those used by the other backends.