Resolve the process crashes when doing key or object folds
Fixes: bz1255

To resolve the process crashes when doing key or object folds, the
Innostore connection port is also saved in the process dictionary
for reuse in asynchronous workers. Without this it is useless to
save the keystore information because using it with a different
port results in a bad argument error when the port_control
function is called.

Remove deps traversal since there are no deps for innostore